+++ CHANGELOG.md
@@ -5,8 +5,12 @@ Unreleased
 ----------
 ### Changed
 * Disabled color-contrast axe-core rule for clinical-lowlight-theme wdio testing
+<<<<<<< HEAD
 ### Added
 * Wdio File Reporter that logs wdio test output to separate json files based on locale, theme and form-factor.
+=======
+* Added `TERRA_THEME_CONFIG` global to the webpack bundle.
+>>>>>>> master
 
 6.0.0 - (April 21, 2020)
 ----------